requirements.txt
Flask
pymongo
python-dotenv
numpy